+++ components/sbm-core/src/main/java/org/springframework/sbm/java/impl/RewriteJavaParser.java
@@ -22,9 +22,8 @@ import org.openrewrite.internal.lang.Nullable;
 import org.openrewrite.java.JavaParser;
 import org.openrewrite.java.marker.JavaSourceSet;
 import org.openrewrite.java.tree.J;
-import org.springframework.sbm.engine.annotations.StatefulComponent;
 import org.springframework.sbm.project.resource.SbmApplicationProperties;
-import org.springframework.sbm.scopeplayground.annotations.ScanScope;
+import org.springframework.sbm.scopes.annotations.ScanScope;
 import org.springframework.stereotype.Component;
 
 import java.nio.file.Path;
